+++ src/components/AsmDB/encodingSpageOverviewPage.tsx
@@ -1,0 +1,3 @@
+export default function EncodingSpaceOverviewPage({ data }: { data: AsmDBData }): JSX.Element {
+  return <p>TODO</p>
+}
