+++ src/components/AsmDB/insnExplainerPage.tsx
@@ -1,0 +1,3 @@
+export default function InsnExplainerPage({ data }: { data: AsmDBData }): JSX.Element {
+  return <p>TODO</p>
+}
